consolidate app shell and nav
