+++ web-app/src/app/app-shell/types.ts
@@ -1,3 +1,9 @@
-export type Config = {
-  GOOGLE_API_KEY: string;
-};
+import { Env, UserPreferences } from "@windy-civi/domain/types";
+
+export interface AppShellLoaderData {
+  env: Env;
+  availableFeeds: string[];
+  preferences: UserPreferences;
+  // lastVisited: string; // timestamp
+  // current
+}
