+++ web-app/src/app/feed/element.tsx
@@ -5,7 +5,6 @@ import { FeedLoaderData } from "./types";
 
 export function Feed() {
   const result = useLoaderData() as FeedLoaderData;
-  console.log("result", result);
 
   return <FeedBills feedData={result.feedData} />;
 }
