--- web-app/src/app/navigator/types.ts
@@ -1,9 +1,0 @@
-import { Env, UserPreferences } from "@windy-civi/domain/types";
-
-export interface NavigatorLoaderData {
-  env: Env;
-  availableFeeds: string[];
-  preferences: UserPreferences;
-  // lastVisited: string; // timestamp
-  // current
-}
