--- web-app/vite.config.d.ts
@@ -1,2 +1,0 @@
-declare const _default: import("vite").UserConfig;
-export default _default;
